requirements.txt
appdirs==1.4.4
asgiref==3.3.4
boto==2.49.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
distlib==0.3.1
Django==3.1.5
django-markwhat==1.6.2
filelock==3.0.12
idna==2.10
keyring==5.7.1
markdown2==2.4.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.5
virtualenv==20.4.7
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
requests==2.25.1
whitenoise==5.2.0
dj_database_url==0.5.0